requirements.txt
alembic==1.9.4
aiohttp==3.8.3
asyncpg==0.27.0
bolt-python==0.7.1
cloud-sql-python-connector==1.2.0
db-dtypes==1.0.5
fastapi==0.89.1
google-auth==2.16.0
google-cloud-bigquery==3.5.0
gunicorn==20.1.0
openai==0.26.5
pandas==1.5.3
sentence-transformers==1.1.0
slack_bolt==1.16.1
sqlalchemy==2.0.4
uvicorn==0.20.0
